+++ hooks_help.c
@@ -6,7 +6,7 @@
 /*   By: mwilk <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/05/08 19:03:53 by mwilk             #+#    #+#             */
-/*   Updated: 2015/11/29 12:07:49 by mwilk            ###   ########.fr       */
+/*   Updated: 2015/12/11 22:52:05 by mwilk            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 void	update_zoom(t_data *d)
 {
-	float c0x;
-	float c0y;
+	double c0x;
+	double c0y;
 
 	c0x = (d->c0x + d->os_x) / d->zoom;
 	c0y = (d->c0y + d->os_y) / d->zoom;
 	d->os_zoom_x += c0x - d->mouse_x * c0x * 2.0;
 	d->os_zoom_y += c0y - d->mouse_y * c0y * 2.0;
-	d->zoom_w = d->zoom * (float)X_HALF;
-	d->zoom_h = d->zoom * (float)Y_HALF;
+	d->zoom_w = d->zoom * (double)X_HALF;
+	d->zoom_h = d->zoom * (double)Y_HALF;
 }
 
 void	hook_key_moves(t_data *d, int keycode)
